Gateway-3389 Refactored for passthrough
