+++ Product/Production/Services/AdminDistributionCore/src/main/java/gov/hhs/fha/nhinc/admindistribution/inbound/AbstractInboundAdminDistribution.java
@@ -26,20 +26,29 @@
  */
 package gov.hhs.fha.nhinc.admindistribution.inbound;
 
+import org.apache.log4j.Logger;
+
 import gov.hhs.fha.nhinc.admindistribution.AdminDistributionAuditLogger;
+import gov.hhs.fha.nhinc.admindistribution.AdminDistributionUtils;
+import gov.hhs.fha.nhinc.admindistribution.adapter.proxy.AdapterAdminDistributionProxy;
+import gov.hhs.fha.nhinc.admindistribution.adapter.proxy.AdapterAdminDistributionProxyObjectFactory;
 import gov.hhs.fha.nhinc.admindistribution.aspect.EDXLDistributionEventDescriptionBuilder;
 import gov.hhs.fha.nhinc.aspect.InboundProcessingEvent;
 import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
 import gov.hhs.fha.nhinc.event.DefaultEventDescriptionBuilder;
+import gov.hhs.fha.nhinc.largefile.LargePayloadException;
 import gov.hhs.fha.nhinc.nhinclib.NhincConstants;
 import oasis.names.tc.emergency.edxl.de._1.EDXLDistribution;
 
 public abstract class AbstractInboundAdminDistribution implements InboundAdminDistribution {
 
-    abstract void processAdminDistribution(EDXLDistribution body, AssertionType assertion);
-
-    protected AdminDistributionAuditLogger auditLogger = new AdminDistributionAuditLogger();
-
+	private static final Logger LOG = Logger.getLogger(AbstractInboundAdminDistribution.class);
+	protected AdminDistributionAuditLogger auditLogger = new AdminDistributionAuditLogger();
+	protected AdminDistributionUtils adminUtils = AdminDistributionUtils.getInstance();
+    protected AdapterAdminDistributionProxyObjectFactory adapterFactory = new AdapterAdminDistributionProxyObjectFactory();
+	
+	abstract void processAdminDistribution(EDXLDistribution body, AssertionType assertion);
+    
     /**
      * This method sends sendAlertMessage to agency/agencies.
      * 
@@ -61,5 +70,16 @@ public abstract class AbstractInboundAdminDistribution implements InboundAdminDi
     private void auditRequestFromNhin(EDXLDistribution body, AssertionType assertion) {
         auditLogger.auditNhinAdminDist(body, assertion, NhincConstants.AUDIT_LOG_INBOUND_DIRECTION,
                 null, NhincConstants.AUDIT_LOG_NHIN_INTERFACE);
+    }
+    
+    protected void sendToAdapter(EDXLDistribution body, AssertionType assertion,
+    		AdminDistributionUtils adminUtils, AdapterAdminDistributionProxyObjectFactory adapterFactory) {
+    	try {
+        	adminUtils.convertDataToFileLocationIfEnabled(body);
+        	AdapterAdminDistributionProxy adapterProxy = adapterFactory.getAdapterAdminDistProxy();
+        	adapterProxy.sendAlertMessage(body, assertion);
+    	} catch (LargePayloadException lpe) {
+            LOG.error("Failed to retrieve payload document.", lpe);
+        }
     }
 }
