+++ Product/Production/Services/AdminDistributionCore/src/test/java/gov/hhs/fha/nhinc/admindistribution/inbound/PassthroughInboundAdminDistributionTest.java
@@ -26,9 +26,11 @@
  */
 package gov.hhs.fha.nhinc.admindistribution.inbound;
 
+import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import gov.hhs.fha.nhinc.admindistribution.AdminDistributionAuditLogger;
@@ -36,6 +38,7 @@ import gov.hhs.fha.nhinc.admindistribution.AdminDistributionUtils;
 import gov.hhs.fha.nhinc.admindistribution.adapter.proxy.AdapterAdminDistributionProxy;
 import gov.hhs.fha.nhinc.admindistribution.adapter.proxy.AdapterAdminDistributionProxyObjectFactory;
 import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
+import gov.hhs.fha.nhinc.common.nhinccommon.NhinTargetSystemType;
 import gov.hhs.fha.nhinc.largefile.LargePayloadException;
 import oasis.names.tc.emergency.edxl.de._1.EDXLDistribution;
 
@@ -55,7 +58,7 @@ public class PassthroughInboundAdminDistributionTest {
         AdminDistributionUtils adminUtils = mock(AdminDistributionUtils.class);
         AdapterAdminDistributionProxyObjectFactory adapterFactory = mock(AdapterAdminDistributionProxyObjectFactory.class);
         AdapterAdminDistributionProxy adapterProxy = mock(AdapterAdminDistributionProxy.class);
-        AdminDistributionAuditLogger auditLogger = new AdminDistributionAuditLogger();
+        AdminDistributionAuditLogger auditLogger = mock(AdminDistributionAuditLogger.class);
         
         when(adapterFactory.getAdapterAdminDistProxy()).thenReturn(adapterProxy);
 
@@ -65,6 +68,10 @@ public class PassthroughInboundAdminDistributionTest {
         passthroughAdminDist.sendAlertMessage(request, assertion);
 
         verify(adapterProxy).sendAlertMessage(eq(request), eq(assertion));
+        
+        verify(auditLogger, times(1)).auditNhinAdminDist(any(EDXLDistribution.class),
+        		any(AssertionType.class), any(String.class), any(NhinTargetSystemType.class), 
+        		any(String.class));
         
     }
